Clear up ambiguity in the intial rotation example
When a user is seeing transform.basis for the first time, it's very important to label all the inputs. People reading this are being presented with several values at once and must understand all of them to understand what is happening in the example. The reader is likely unfamiliar with what `transform.basis` actually does. They are also given a Vector3 without explanation that it is the axis, they are given PI without explanation that that is the amount of rotation, they are multiplying by the value of the transform.basis again hinging on them understanding what that is. Finally, they are creating a Basis object and it's not made clear that the two values this receives are actually an axis and an amount.

Rotating by PI makes the rotation direction very unclear. It could only be worse if we rotated by TAU. 

I'm introducing the changes above to hopefully rectify a lot of these issues and make things easier for the next person.